chore(lint): make pyproject the single source of ruff truth (#107)
The two ruff pre-commit hooks each passed an explicit --select, which
overrode [tool.ruff.lint] in pyproject.toml. CI therefore enforced only
D, PD, W505, UP, F401, F811 and I, while `ruff check` run by hand used
the much broader pyproject set. The two disagreed by 123 findings.

Collapse both hooks into one `ruff-check` with no --select, so the
config in pyproject.toml is what runs everywhere. The redundant
--line-length and --config args go too, and the deprecated `ruff`
alias becomes `ruff-check`. PD moves into pyproject's select so the
enforced rule set is a strict superset of what the hooks ran before.

Pin black to line-length 80. Nothing configured black, so it reflowed
to its default 88 while ruff's E501 wanted 80 -- the source of 64 of
the 123 findings. This reformats 20 files; no code changes meaning.

Real fixes for the rest:
- drop 48 dead noqa directives (46 SLF001, plus BLE001) for rules that
  are not in the select list
- RUF043: mark the pytest.raises match= pattern raw
- RUF046: drop an int() around a single-arg round(), already an int

Rules the codebase deliberately does not follow move to per-file
ignores with the reason stated, replacing scattered suppressions:
in-test imports (PLC0415), benchmark import-after-importorskip and
long measurement sweeps (E402, PLR0912, PLR0915), the private
re-exports in transport/__init__.py (F401), and N818 on
ConnectionLost, which is exported from the package root and so cannot
be renamed without breaking published API.

Five noqa directives that RUF100 reported as unused are in fact
load-bearing; removing them surfaces nine real findings. They are
resolved via the per-file ignores above rather than deleted.

+++ tests/benchmark/tinyros/test_speed_interprocess.py
@@ -149,7 +149,9 @@ def _build_bytes(i: int, n_bytes: int) -> bytes:
     return prefix + (b"x" * pad)
 
 
-def _build_ndarray(i: int, shape: tuple[int, ...], dtype: np.dtype) -> np.ndarray:
+def _build_ndarray(
+    i: int, shape: tuple[int, ...], dtype: np.dtype
+) -> np.ndarray:
     """Build a zero ndarray with ``flat[0] = i``.
 
     Args:
@@ -264,7 +266,7 @@ def _percentile(values: list[float], q: float) -> float:
     """
     if not values:
         return 0.0
-    k = max(0, min(len(values) - 1, int(round(q * (len(values) - 1)))))
+    k = max(0, min(len(values) - 1, round(q * (len(values) - 1))))
     return sorted(values)[k]
 
 
@@ -431,7 +433,8 @@ def _run_case(
                 got_idx == i
             ), f"[{label}] step {step}: expected idx {i}, got {got_idx}"
             assert got_count == i + 1, (
-                f"[{label}] step {step}: expected count {i + 1}, " f"got {got_count}"
+                f"[{label}] step {step}: expected count {i + 1}, "
+                f"got {got_count}"
             )
             delivered += 1
             if SLEEP_BETWEEN_ITERS_S > 0:
@@ -517,4 +520,6 @@ def test_interprocess_latency_matrix() -> None:
             delivered == REPETITIONS
         ), f"[{label}] delivery mismatch: got {delivered}/{REPETITIONS}"
     _save_csv(rows, CSV_DIR / "interprocess_latency.csv")
-    assert len(rows) == len(_PAYLOADS), f"expected one row per payload, got {len(rows)}"
+    assert len(rows) == len(
+        _PAYLOADS
+    ), f"expected one row per payload, got {len(rows)}"
